升级了nutz、druid的版本 优化了频高图扫描速度慢得问题，现在扫描胶片频高图一百多万耗时大概在4小时内处理结束

+++ src/cn/fam1452/action/ht/PgtMod.java
@@ -774,8 +774,10 @@ public class PgtMod extends BaseMod{
 						iglist.add(ig) ;
 						ndyYear.add(DateUtil.getYearstrByDate(ig.getCreateDate())) ;
 						//log.info( ig.getGramPath() ) ;
-						baseService.dao.delete(ig) ;
-						baseService.dao.insert(ig) ;
+						if(baseService.dao.count(IronoGram.class, Cnd.where("gramID", "=", ig.getGramID())) > 0){
+							baseService.dao.delete(ig) ;
+						}
+						baseService.dao.fastInsert(ig) ;
 					}else{
 						fail++ ;
 						failFile.append(fn).append(",") ;
